# backend/static/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/organizations/{orgId}/projects/{projectId}/issues": {
      "post": {
        "summary": "Create a new issue",
        "tags": [
          "Issue"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "createIssue",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "urgent",
                      "high",
                      "medium",
                      "low",
                      "none"
                    ]
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "task",
                      "bug",
                      "story",
                      "epic"
                    ]
                  },
                  "assignee_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "parent_issue_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "due_date": {
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201",
            "content": {
              "application/json": {
                "schema": {
                  "0": "I",
                  "1": "s",
                  "2": "s",
                  "3": "u",
                  "4": "e",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "project_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "issue_key": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "backlog",
                        "todo",
                        "in_progress",
                        "in_review",
                        "in_dev",
                        "in_prod",
                        "cancelled"
                      ]
                    },
                    "priority": {
                      "type": "string",
                      "enum": [
                        "urgent",
                        "high",
                        "medium",
                        "low",
                        "none"
                      ]
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "task",
                        "bug",
                        "story",
                        "epic"
                      ]
                    },
                    "assignee_id": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "reporter_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "sort_order": {
                      "type": "number"
                    },
                    "parent_issue_id": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "due_date": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "project_id",
                    "issue_key",
                    "title",
                    "description",
                    "status",
                    "priority",
                    "type",
                    "assignee_id",
                    "reporter_id",
                    "sort_order",
                    "parent_issue_id",
                    "due_date",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "List issues for a project with optional filters",
        "tags": [
          "Issue"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "operationId": "listIssues",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "0": "I",
                        "1": "s",
                        "2": "s",
                        "3": "u",
                        "4": "e",
                        "5": "W",
                        "6": "i",
                        "7": "t",
                        "8": "h",
                        "9": "D",
                        "10": "e",
                        "11": "t",
                        "12": "a",
                        "13": "i",
                        "14": "l",
                        "15": "s",
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "project_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "issue_key": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "backlog",
                              "todo",
                              "in_progress",
                              "in_review",
                              "in_dev",
                              "in_prod",
                              "cancelled"
                            ]
                          },
                          "priority": {
                            "type": "string",
                            "enum": [
                              "urgent",
                              "high",
                              "medium",
                              "low",
                              "none"
                            ]
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "task",
                              "bug",
                              "story",
                              "epic"
                            ]
                          },
                          "assignee_id": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "reporter_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "sort_order": {
                            "type": "number"
                          },
                          "parent_issue_id": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "due_date": {
                            "type": "string",
                            "nullable": true
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          },
                          "assignee": {
                            "0": "U",
                            "1": "s",
                            "2": "e",
                            "3": "r",
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "clerk_user_id": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string",
                                "format": "email"
                              },
                              "first_name": {
                                "type": "string",
                                "nullable": true
                              },
                              "last_name": {
                                "type": "string",
                                "nullable": true
                              },
                              "avatar_url": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                              },
                              "is_active": {
                                "type": "boolean"
                              },
                              "last_login_at": {
                                "type": "string",
                                "nullable": true
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "updated_at": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "clerk_user_id",
                              "email",
                              "first_name",
                              "last_name",
                              "avatar_url",
                              "is_active",
                              "last_login_at",
                              "created_at",
                              "updated_at"
                            ],
                            "nullable": true
                          },
                          "reporter": {
                            "0": "U",
                            "1": "s",
                            "2": "e",
                            "3": "r",
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "clerk_user_id": {
                                "type": "string"
                              },
                              "email": {
                                "type": "string",
                                "format": "email"
                              },
                              "first_name": {
                                "type": "string",
                                "nullable": true
                              },
                              "last_name": {
                                "type": "string",
                                "nullable": true
                              },
                              "avatar_url": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                              },
                              "is_active": {
                                "type": "boolean"
                              },
                              "last_login_at": {
                                "type": "string",
                                "nullable": true
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "updated_at": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "clerk_user_id",
                              "email",
                              "first_name",
                              "last_name",
                              "avatar_url",
                              "is_active",
                              "last_login_at",
                              "created_at",
                              "updated_at"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "project_id",
                          "issue_key",
                          "title",
                          "description",
                          "status",
                          "priority",
                          "type",
                          "assignee_id",
                          "reporter_id",
                          "sort_order",
                          "parent_issue_id",
                          "due_date",
                          "created_at",
                          "updated_at"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/organizations/{orgId}/projects/{projectId}/issues/{issueId}/assign": {
      "patch": {
        "summary": "Assign or unassign an issue",
        "tags": [
          "Issue"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "issueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "operationId": "assignIssue",
        "requestBody": {
          "description": "Body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assignee_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                },
                "required": [
                  "assignee_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "0": "I",
                  "1": "s",
                  "2": "s",
                  "3": "u",
                  "4": "e",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "project_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "issue_key": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "backlog",
                        "todo",
                        "in_progress",
                        "in_review",
                        "in_dev",
                        "in_prod",
                        "cancelled"
                      ]
                    },
                    "priority": {
                      "type": "string",
                      "enum": [
                        "urgent",
                        "high",
                        "medium",
                        "low",
                        "none"
                      ]
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "task",
                        "bug",
                        "story",
                        "epic"
                      ]
                    },
                    "assignee_id": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "reporter_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "sort_order": {
                      "type": "number"
                    },
                    "parent_issue_id": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "due_date": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "project_id",
                    "issue_key",
                    "title",
                    "description",
                    "status",
                    "priority",
                    "type",
                    "assignee_id",
                    "reporter_id",
                    "sort_order",
                    "parent_issue_id",
                    "due_date",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "version": "1.0.0",
    "title": "Vector REST API - Documentation",
    "description": "Vector REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}